Adds initial named unit systems functionaliity
The named_system.* files have the basic code
for storing the unit system names. These are
stored with an index that can be used in
conjunction with a bitmap that can be built
into the units to mark which systems they
belong to.

The testUnits-2.c file has the unit-tests
for the added code.

ut_free_system and the Makefile are updated
to free up memory and include the
appropriate files.