Adds utility structure ut_string_list to the lib.
This includes a readonly public API for library
users to use (declared in udunits2.h) and an
internal/private API (declared in utlists.h) for
creating, building and manipulating the string lists.

This should likely be reworked to be generic arrays
of void* pointers that then could be strings or ut_units
or anything. If could also have the private API extended
significantly.

With the current implementation NULL pointers can be
stored in the lists.

Changes committed:
	new file:   lib/ut_lists.c
	new file:   lib/ut_lists.h