Functional named units system API added.
This has all the library interfaces for managing
the named units systems part of the library. It has
a unit test suite for the added structures and
functions.

It lacks integration into the XML parsing to assign
units to named groups from the starte when loaking from
XML.

And it lacks the ability to ask for ut_units that
are convertible from other units.