Implements named unit systems in the XML parser.
This adds the ability to specify the named system(s)
a unit belongs to as attributes in the XML file. The
attribute name is "system" or "systems" ant is set to
a comma separated list of system names. For example
this would be <unit systems="SI, metric">. Leading
and trailing spaces are trimmed from system names and
comparisons are case-insensitive. The code added to
xml.c is ifdef'd with ALLOW_NAMED_UNIT_SYSTEMS. Just
uncomment the undef to remove the added capabilites.

Test cases and an XML file were added.

Still needed is the way to defince aliases, etc., for
named unit systems using XML.

Changes committed:
	modified:   lib/testUnits-2.c
	modified:   lib/xml.c
	modified:   lib/xmlSuccesses/Makefile.am
	new file:   lib/xmlSuccesses/named_systems.xml