Optional automatic creation of named unit systems
Make it a compile time option for adding a unit to
a named unit system or group to automatically create
the group if it does not exists.

Also implements a similar feature for loading XML
files except it is less optional at this time
since pre-creating them is not yet possible in XML.

Both of these could be switched to run-time options
by creating appropriate global or status vaials.