Makes named-unit-system auto creation runtime alterable
An interface is needed to swith the option on and
off but whether non-existent named unit systems are
automatically created in a <unit systems="a,b,c">
tag is controlled by the static variable:
  __auto_create_named_systems_in_xml

Having this set to true is simplest since the system
names can just be defined on the fly in the <unit>
tag attributes but this eliminates the possibility
of any error checking.

Having this set to false means before any system
names can be used in <unit> attributes they must
be defined in a <named-unit-system> tag. This
makes it a parse error if there is a typo in the
<unit> attribute systems list.